requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pandas
openpyxl
pypdf
python-dotenvfastapi
uvicorn
python-multipart
google-generativeai
pandas
openpyxl
pypdf
python-dotenv